requirements.txt
opencv-python
dlib
imutils
numpy
Flask
Pillow